requirements.txt
python-vxi11>=0.9
PyYAML>=6.0
